Add dotfiles submodule that is the home-directory skeleton for users.
It is desired that there be a non-trivial system-wide home-directory skeleton
for all users.  As a separate repository, it is cloned into a user's $HOME (by
itself, without the superproject), to populate their instance of the skeleton so
that changes can be pulled/pushed between users' clones and the system-wide
upstream.  As a submodule at /etc/nixos/users/dotfiles, related changes across
it and the superproject at /etc/nixos can be kept synchronized.  The dotfiles
primary repository's $GIT_DIR is kept in the superproject's primary
.git/modules/users/dotfiles, and so this sub-directory structure should be
preserved, which corresponds to an intended basic method of deployment in new
NixOS installations of just copying the /etc/nixos/ contents into place (not
cloning).  It is intended for both repositories to be able to not have any
upstream, when located at /etc/nixos.