Set GIT_ASKPASS="" and use ssh-askpass-fullscreen.
To prevent Git from using SSH_ASKPASS (which NixOS always sets), and to avoid
x11-ssh-askpass which is buggy in that it sometimes does not show its GUI window
with MATE Desktop at least.

This commit undoes the previous commit and so re-allows SSH_ASKPASS to be set,
which is desirable for SSH uses that do not have a terminal.