Configure mDNS and LLMNR consistently, by my.nameResolv.multicast.
Only resolving is enabled for them, not responding.  But also add my.publish
options that give the ability to enable responding or to disable resolving.

Add firewall logic to open ports as needed for the mDNS and LLMNR modes, and to
assert that only those intended are actually opened.

Might not be perfect yet but it's a start of something better.