Set enableDebugInfo=true for GDB debugging, for host yoyo only.
If the size of these grows too big in the future, then should be able to remove
this (so it defaults to false) and instead directly install the "debug" output
of particular packages that are really desired, e.g. glibc.debug.  That could be
done per-user by Home Manager instead, even.