Ensure that a local user name is used for commits to their ~/.dotfiles.
Remove this step from the Install Guide, since it's now always done
automatically.

I want this change, so that it's done also for users that are created after the
system is installed.  Otherwise, when creating such a new user later, I forgot
and had commits in their ~/.dotfiles that had an undesired author value (because
their ~/.config/git/config uses a different value intended for non-dotfiles
files).