Fix prepare-installation's setup of the nixos-unstable channel.
The previous code worked with an older NixOS version's live installer,
but it stopped working with 22.05's, IIRC, and it definitely does not
work with 22.11's.

This fix also makes the logic of my installation steps a little
cleaner for this aspect of setting-up the nixos-unstable channel in
the installee machine before it is ever booted.