Fix /etc/machine-id to exist directly in the / FS, for systemd boot.
Previously, I'd placed that file in /state/etc/machine-id and symlink'ed /etc/machine-id to that.
IIRC, I'd done this following the older version of the "NixOS Root on ZFS" guide.  It used to work
it seemed, but at some point after an upgrade to a new NixOS release version with a new SystemD
version, this broke, many months ago.  This was preventing persisting the systemd journal to
/var/log/journal/.