Adjustments for upgrade to NixOS 23.11.
The Tuxedo Control Center had to be removed because the needed old version of NodeJS is no longer
provided by Nixpkgs (because it's end-of-life), and because TCC was never accepted into Nixpkgs,
and because the separate Nix package of it has recently become unmaintained.  Until an alternative
is setup, I won't be able to change the performance profile of my CPU, GPU, and fans.

The TrustCor certificates can't be blacklisted anymore because those exact names are no longer in
the default set of certificates, and it seems that there aren't any from TrustCor anymore which is
good.