Fix for outside bug with GRUB2 and OpenZFS 2.2+ boot zpools.
When a snapshot is taken in a zpool, the `snapshots_changed` property is (automatically and
unavoidably) added to the dataset (which seems to be a new thing introduced in OpenZFS 2.2).  This
property uses and activates the `extensible_dataset` zpool feature if that feature was
available (either `enabled` or `active`).  That feature and others which use it are included in
the stock `grub2` "compatibility feature set" file (and have been for a long time) and thus it was
`enabled` for boot zpools.  However, that feature is actually not supported by GRUB2, and if it's
`active` then the boot zpool cannot be booted from.  I was unable to boot into my system because
of this, after I simply took a snapshot of my boot zpool's datasets including the top
dataset (because now with OpenZFS 2.2, after upgrading to NixOS 23.11, taking a snapshot added the
new `snapshots_changed` property which activated the `extensible_dataset` feature which previously
was not `active` (it was only `enabled`)).  Once this feature is activated it cannot be undone,
and thus the boot zpool can never again be booted from.

I had to destroy that zpool and recreate it (hopefully exactly the same) via `zfs receive` from a
`zfs send -R` backup stream of the original.  The new boot zpool was created with my new custom
"compatibility feature set" file which ensures that feature is and will continue to be disabled,
so when that property is added (when a snapshot is taken) it cannot activate that feature (and
thus that property is not persistent across unmounts, which I don't care about especially not for
the boot zpool).  (In contrast, for my main zpool, which was unharmed, that property is persistent
since that feature is now activated in it.)

To ensure this doesn't happen to me again, my `prepare-installation` script is also updated to
create boot zpools using this more-restricted feature set.

My new custom "compatibility feature set" file is now installed in my systems, so it remains
available for future ZFS operations such as `zpool upgrade` which will need to find it since it's
now referenced by my boot zpool's `compatibility` property.