Change nix-daemon tasks' scheduling policy to "idle", for host yoyo.
So that when I do nixos-rebuild, and it usually takes a while, and when I also do other tasks
simultaneously (e.g. login to another user, unlock KeePass DB, start web browser), those other
tasks will be given precedence for CPU and I/O so they're more responsive, at the cost of the Nix
builds being a little delayed.

This might not be wanted for all hosts, so I place these options' definitions in the per-host
module.