Install KeePassXC, use as Secret Service. Improve handling of creds.
There is a corresponding commit in my dotfiles repo that:
configures KeePassXC to provide the Secret Service API (as needed by this commit), and
configures it to integrate with the ssh-agent (which is now enabled by this commit), and
configures its other features to my liking.

My customization of the Firefox package now allows the KeePassXC-Browser extension to do
native-messaging with KeePassXC (via its `keepassxc-proxy` executable), which is needed when a
user installs that extension.  Whether or not the `keepassxc` package is installed in
`environment.systemPackages`, this will still pull in the `keepassxc` package in the NixOS system
closure, but that alone doesn't expose its executables in PATH.

The `myFirefox` package is removed and no longer installed in `environment.systemPackages`,
because that was actually pointless because my users install Firefox via Home Manager, which
overrode the system-wide one.  Now, my customization of the `firefox` package is an overlay so
that it takes effect everywhere including for users installing it via Home Manager.  This is
needed for users to have the native-messaging for the browser extension.

Unfortunately, for (Ungoogled) Chromium, there doesn't seem to be a way to achieve system-wide
allowing of the native-messaging for the browser extension.  Having that is needed for the
extension to work in Chromium, and so the corresponding commit in my dotfiles repo arranges for
Home Manager to setup that, which is a less-clean way but at least is automatically provided for
users who want to use Ungoogled Chromium.  Because Chromium is not my primary browser, I'm less
concerned with this.