Have setup-home dump Home Manager news so it'll be quiet about this.
Because this script is usually used to setup new users in which case it's usually undesired to see
all of Home Manager's changes as "what's new".  Its news that's actually new will still be
mentioned when it's updated in the future.  Previously, I'd commented this out, but have changed
my mind.