Fix not-having compression across the entire main pool.
It was changed in OpenZFS 2.2 from a default of `compression=off` to `compression=on` [1].  When
I'd first made the `prepare-installation` script w/o giving a value for that property, it wasn't
being enabled by default.

Unfortunately, upgrading to OpenZFS 2.2 as part of upgrading the NixOS version (many months ago),
caused `compression=on` to become enabled for preexisting installations where it was off
before (and to be enabled for new installations, of course).  I think I have been noticing some
things being slower since (on my same exact computer and drives).

It can be changed back to off across the entire main pool by simply doing:
    root# zfs set compression=off $MAIN_POOL
But all already-compressed files (and zvol blocks) that were compressed when it was on will remain
so.  Fortunately, with NixOS often switching to different files for installed packages, many of
those compressed files will be phased-out and the replacements will not get compressed anymore.
Those package files seem to be a significant fraction of what affects my performance.  But other
compressed files that aren't replaced will remain so (unless you identify and replace them).

I guess I should've explicitly set it to off to ensure that across version changes, otherwise it
seems that OpenZFS assumes they can change defaults and we'll be alright with that.  Annoyingly,
their release notes for 2.2 didn't even mention the change [2].

[1] https://github.com/openzfs/zfs/commit/56fa4aa96eb3875f254e93eaef646ea20ba187f9
[2] https://github.com/openzfs/zfs/releases/tag/zfs-2.2.0